Phishing Domain Detection/requirements.txt
numpy
re
requests
tldextract
whois
joblib
socket
streamlit as st
datetime
time
dns.resolver
urllib.parse
warnings